+++ addon/transforms/timestamp.js
@@ -1,3 +1,5 @@
+/* global Firebase */
+
 import Transform from './base';
 
 export default Transform.extend({
@@ -9,6 +11,9 @@ export default Transform.extend({
   },
 
   serialize: function(value) {
+    if (value === Firebase.ServerValue.TIMESTAMP) {
+      return value;
+    }
     if (!value || !value.getTime) {
       return null;
     }